+++ knative-operator/pkg/webhook/util/util.go
@@ -3,10 +3,15 @@ package util
 import (
 	"context"
 	"fmt"
+	"github.com/appscode/jsonpatch"
 	"github.com/coreos/go-semver/semver"
 	configv1 "github.com/openshift/api/config/v1"
+	admissionv1beta1 "k8s.io/api/admission/v1beta1"
+	"net/http"
 	"os"
 	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
+	"sigs.k8s.io/controller-runtime/pkg/webhook/admission/types"
 )
 
 // ValidateOpenShiftVersion validates the current openshift version against the minimum
@@ -40,4 +45,21 @@ func ValidateOpenShiftVersion(ctx context.Context, cl client.Client) (bool, stri
 		return false, msg, nil
 	}
 	return true, "", nil
+}
+
+// PatchResponseFromRaw takes 2 byte arrays and returns a new response with json patch.
+// The original object should be passed in as raw bytes to avoid the roundtripping problem
+// described in https://github.com/kubernetes-sigs/kubebuilder/issues/510.
+func PatchResponseFromRaw(original, current []byte) types.Response {
+	patches, err := jsonpatch.CreatePatch(original, current)
+	if err != nil {
+		return admission.ErrorResponse(http.StatusInternalServerError, err)
+	}
+	return types.Response{
+		Patches: patches,
+		Response: &admissionv1beta1.AdmissionResponse{
+			Allowed:   true,
+			PatchType: func() *admissionv1beta1.PatchType { pt := admissionv1beta1.PatchTypeJSONPatch; return &pt }(),
+		},
+	}
 }
